drivers: input: sec_input: introduce support for Samsung Dex touchpad
It would be more better if the commit message would be 'drivers: input: sec_input: enable support for Samsung Dex touchpad' as we already had support for it, the main issue it wasn't enabled, and it is normal as the stock rom doesn't have and kind of support for Samsung Dex, so yes this commit hard codes the value in kernel instead of depending on dtbo. I think it is much better instead of playing directly into dtbo/dtb although the original way to do this is based on dtbo but anyway.

 * also enables support for ear detect but useless i think?

Signed-off-by: nostalgiceagle <183935403+nostalgiceagle@users.noreply.github.com>